+++ test/support/worker_config.ex
@@ -2,4 +2,5 @@ defmodule WorkerConfig do
   use Faktory.Configuration, :worker
 
   port 7421
+  pool 5
 end